Use common definitions for all the TestSpans(SpansCatalog) and _NetCDFVarWithDimensions.
