+++ lib/iris/tests/unit/fileformats/cf/test_CFBoundaryVariable.py
@@ -6,25 +6,10 @@
 
 from iris.fileformats.cf import CFBoundaryVariable
 
-from .identify_catalogue import IdentifyByAttributeCatalog
-
-
-class _NetCDFVarWithDimensions:
-    """Stub with a dimensions attribute for spans() tests."""
-
-    def __init__(self, name, dimensions, dtype=int):
-        import numpy as np
-
-        self.name = name
-        self.dtype = np.dtype(dtype)
-        self.dimensions = dimensions
-
-    def ncattrs(self):
-        return [
-            attr
-            for attr in self.__dict__
-            if not attr.startswith("_") and attr not in ["name", "dtype", "dimensions"]
-        ]
+from .identify_catalogue import (
+    IdentifyByAttributeCatalog,
+    SpansCatalog,
+)
 
 
 class TestIdentify(IdentifyByAttributeCatalog):
@@ -52,35 +37,8 @@ class TestIdentify(IdentifyByAttributeCatalog):
         assert expected == result
 
 
-class TestSpans:
+class TestSpans(SpansCatalog):
     """Tests for CFBoundaryVariable.spans()."""
 
-    def _make_cf_var(self, name, dimensions):
-        stub = _NetCDFVarWithDimensions(name, dimensions)
-        return CFBoundaryVariable(name, stub)
-
-    def test_empty_dimensions_spans(self):
-        """Scalar boundary variable always spans the target."""
-        cf_boundary = self._make_cf_var("bounds_var", ())
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_boundary.spans(cf_target)
-
-    def test_source_trailing_subset_spans(self):
-        """source[:-1] is a subset of target dimensions => spans."""
-        # bounds_var has dims (x, y, bounds_extent); data_var has (x, y)
-        cf_boundary = self._make_cf_var("bounds_var", ("x", "y", "bounds_extent"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_boundary.spans(cf_target)
-
-    def test_source_leading_subset_spans(self):
-        """source[1:] is a subset of target dimensions => spans."""
-        # bounds_var has dims (bounds_extent, x, y); data_var has (x, y)
-        cf_boundary = self._make_cf_var("bounds_var", ("bounds_extent", "x", "y"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_boundary.spans(cf_target)
-
-    def test_non_spanning(self):
-        """Dimensions that don't fit either slice => does not span."""
-        cf_boundary = self._make_cf_var("bounds_var", ("x", "b", "c"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert not cf_boundary.spans(cf_target)
+    __test__ = True
+    CF_CLASS = CFBoundaryVariable
