+++ lib/iris/tests/unit/fileformats/cf/test_CFClimatologyVariable.py
@@ -6,25 +6,10 @@
 
 from iris.fileformats.cf import CFClimatologyVariable
 
-from .identify_catalogue import IdentifyByAttributeCatalog
-
-
-class _NetCDFVarWithDimensions:
-    """Stub with a dimensions attribute for spans() tests."""
-
-    def __init__(self, name, dimensions, dtype=int):
-        import numpy as np
-
-        self.name = name
-        self.dtype = np.dtype(dtype)
-        self.dimensions = dimensions
-
-    def ncattrs(self):
-        return [
-            attr
-            for attr in self.__dict__
-            if not attr.startswith("_") and attr not in ["name", "dtype", "dimensions"]
-        ]
+from .identify_catalogue import (
+    IdentifyByAttributeCatalog,
+    SpansCatalog,
+)
 
 
 class TestIdentify(IdentifyByAttributeCatalog):
@@ -52,33 +37,8 @@ class TestIdentify(IdentifyByAttributeCatalog):
         assert expected == result
 
 
-class TestSpans:
+class TestSpans(SpansCatalog):
     """Tests for CFClimatologyVariable.spans()."""
 
-    def _make_cf_var(self, name, dimensions):
-        stub = _NetCDFVarWithDimensions(name, dimensions)
-        return CFClimatologyVariable(name, stub)
-
-    def test_empty_dimensions_spans(self):
-        """Scalar climatology variable always spans the target."""
-        cf_clim = self._make_cf_var("clim_var", ())
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_clim.spans(cf_target)
-
-    def test_source_trailing_subset_spans(self):
-        """source[:-1] is a subset of target dimensions => spans."""
-        cf_clim = self._make_cf_var("clim_var", ("x", "y", "clim_extent"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_clim.spans(cf_target)
-
-    def test_source_leading_subset_spans(self):
-        """source[1:] is a subset of target dimensions => spans."""
-        cf_clim = self._make_cf_var("clim_var", ("clim_extent", "x", "y"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_clim.spans(cf_target)
-
-    def test_non_spanning(self):
-        """Dimensions that don't fit either slice => does not span."""
-        cf_clim = self._make_cf_var("clim_var", ("x", "b", "c"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert not cf_clim.spans(cf_target)
+    __test__ = True
+    CF_CLASS = CFClimatologyVariable
