+++ lib/iris/tests/unit/fileformats/cf/test_CFLabelVariable.py
@@ -9,23 +9,11 @@ import pytest
 
 from iris.fileformats.cf import CFDataVariable, CFLabelVariable
 
-from .identify_catalogue import IdentifyByAttributeCatalog
-
-
-class _NetCDFVarWithDimensions:
-    """Stub with a dimensions attribute for spans() and cf_label_dimensions() tests."""
-
-    def __init__(self, name, dimensions, dtype=int):
-        self.name = name
-        self.dtype = np.dtype(dtype)
-        self.dimensions = dimensions
-
-    def ncattrs(self):
-        return [
-            attr
-            for attr in self.__dict__
-            if not attr.startswith("_") and attr not in ["name", "dtype", "dimensions"]
-        ]
+from .identify_catalogue import (
+    IdentifyByAttributeCatalog,
+    SpansCatalog,
+    _NetCDFVarWithDimensions,
+)
 
 
 class TestIdentify(IdentifyByAttributeCatalog):
@@ -79,33 +67,8 @@ class TestCfLabelDimensions:
         assert result == ()
 
 
-class TestSpans:
+class TestSpans(SpansCatalog):
     """Tests for CFLabelVariable.spans()."""
 
-    def _make_cf_var(self, name, dimensions):
-        stub = _NetCDFVarWithDimensions(name, dimensions, dtype=np.bytes_)
-        return CFLabelVariable(name, stub)
-
-    def test_empty_dimensions_spans(self):
-        """Scalar label variable always spans the target."""
-        cf_label = self._make_cf_var("label_var", ())
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_label.spans(cf_target)
-
-    def test_source_trailing_subset_spans(self):
-        """source[:-1] (drop string length dim) is subset of target => spans."""
-        cf_label = self._make_cf_var("label_var", ("x", "strlen"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_label.spans(cf_target)
-
-    def test_source_leading_subset_spans(self):
-        """source[1:] is a subset of target dimensions => spans."""
-        cf_label = self._make_cf_var("label_var", ("strlen", "x"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert cf_label.spans(cf_target)
-
-    def test_non_spanning(self):
-        """Dimensions that don't fit either slice => does not span."""
-        cf_label = self._make_cf_var("label_var", ("x", "b", "c"))
-        cf_target = self._make_cf_var("data_var", ("x", "y"))
-        assert not cf_label.spans(cf_target)
+    __test__ = True
+    CF_CLASS = CFLabelVariable
